+++ server/src/asm/parser.ts
@@ -621,7 +621,7 @@ export class Parser extends Tokenizer {
     }
 
     token.type = TokenType.Macro
-    const symExp = this.newSymbolExpression([token], SymbolType.TypeName, false)
+    const symExp = this.newSymbolExpression([token], SymbolType.MacroName, false)
     return this.initStatement(new stm.MacroInvokeStatement(), symExp)
   }
 
@@ -712,6 +712,7 @@ export class Parser extends Tokenizer {
             return
           }
         // TODO: hack for handling indented data defs in .structs
+        // (This will go away once parsing is integrated into pass 0)
         } else if (t2str == ".byte"
             || t2str == ".res"
             || t2str == ".dbyt"